fix(routes): move Sleeper data pages to root paths
